rethink container cleanup (SOFTWARE-4762)
the trap in supervisord_startup.sh never gets run because supervisord
is started with 'exec', which replaces the bash process.  (So, bash will
no longer be around to run the trap code.)

Two ways came to mind to fix this:

1. remove the 'exec' keyword.

   This seems the simplest, but will result in supervisord being PID 2,
   and the initial supervisord_startup.sh process (bash) will stay PID 1

2. move the trap to a highest-prio supervisord service.

   When supervisord finishes, it will stop this 'service' last, which
   should trigger the cleanup code after everything else.

It was favored in SOFTWARE-4762 to keep supervisord at PID 1, so we've
gone with implementation 2 above.